+++ BACKEND/api/routes/cam_modules.py
@@ -12,8 +12,10 @@ router = APIRouter(prefix="/cam_modules", tags=["cam_modules"])
 
 logger = logging.getLogger(f"{settings.app_name}.{__name__}")
 
-face_cc_image_queue: None | ImageQueue = None
-face_yn_image_queue: None | ImageQueue = None
+image_queues: dict[str, ImageQueue | None] = {
+    "face_cc": None,
+    "face_yn": None,
+}
 
 
 @router.websocket("face_cc", name="face_cc")
@@ -22,7 +24,7 @@ async def face_cc(websocket: WebSocket):
     This is the endpoint that we will be sending request to from the
     frontend
     """
-
+    face_cc_image_queue = image_queues.get("face_cc")
     if face_cc_image_queue is None:
         raise HTTPException(
             status_code=500,
@@ -42,28 +44,29 @@ async def face_yn(websocket: WebSocket):
     This is the endpoint that we will be sending request to from the
     frontend
     """
-
-    if face_cc_image_queue is None:
+    face_yn_image_queue = image_queues.get("face_yn")
+    if face_yn_image_queue is None:
         raise HTTPException(
             status_code=500,
-            detail=f"Error not ready face_cc_image_queue.",
+            detail=f"Error not ready face_yn_image_queue.",
         )
 
     try:
-        await face_cc_image_queue.loop(websocket)
+        await face_yn_image_queue.loop(websocket)
     except Exception as e:
         # logger.error(e)
         ...
 
 
 @router.on_event("startup")
-async def startup():
-    global face_cc_image_queue
+async def startup(queues: dict[str, ImageQueue] = None):
     """
     This tells fastapi to load the classifier upon app startup
     so that we don't have to wait for the classifier to be loaded after making a request
     """
+    if queues is None:
+        queues = image_queues
     cc_func = CascadeClassierFun()
     cc_func.load()
-    face_cc_image_queue = ImageQueue(cc_func)
-    face_yn_image_queue = ImageQueue(cc_func)
+    queues["face_cc"] = ImageQueue(cc_func)
+    queues["face_yn"] = ImageQueue(cc_func)
