+++ BACKEND/api/services/plate_detection/image_ops.py
