+++ BACKEND/api/services/plate_detection/models/haarcascade_russian_plate_number.xml
