+++ BACKEND/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "fastparking-api"
-version = "1.6.4"
+version = "1.6.5"
 description = "GoIT DataScience Team Project - Fast Parking API module"
 authors = ["neuron_curva_bobr <info@neuron.curva.bobr.app>"]
 license = "MIT"
